*Some* refactoring
