+++ app/grpc/flow_type_handler.rb
@@ -6,7 +6,7 @@ class FlowTypeHandler < Tucana::Sagittarius::FlowTypeService::Service
   def update(request, _call)
     current_runtime = Runtime.find(Code0::ZeroTrack::Context.current[:runtime][:id])
 
-    response = Runtimes::FlowTypes::UpdateService.new(current_runtime, request.flow_types).execute
+    response = Runtimes::Grpc::FlowTypes::UpdateService.new(current_runtime, request.flow_types).execute
 
     Tucana::Sagittarius::FlowTypeUpdateResponse.new(success: response.success?)
   end
